services/common/requirements/prod.txt
aiohttp
python-dotenv
sentry-sdk[flask]
asyncio-nats-client
asyncio-nats-streaming
